format_to_parts
